# app/Models/Project.php
<?php
namespace App\Models;
use Illuminate\Support\Str;
use Illuminate\Support\Collection;
use Illuminate\Database\Eloquent\Model;
use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\SoftDeletes;

/**
 * @OA\Schema(
 *     schema="Project",
 *     title="Project",
 *     description="Project model",
 *     @OA\Property(property="id", type="string", format="uuid", example="550e8400-e29b-41d4-a716-446655440000"),
 *     @OA\Property(property="organization_id", type="string", format="uuid"),
 *     @OA\Property(property="title", type="string", example="Construction d'un complexe scolaire"),
 *     @OA\Property(property="project_code", type="string", example="PRJ-2026-001"),
 *     @OA\Property(property="status", type="string", enum={"draft", "proposed", "approved", "rejected"}),
 *     @OA\Property(property="start_date", type="string", format="date"),
 *     @OA\Property(property="end_date", type="string", format="date")
 * )
 */
class Project extends Model
{
    use HasFactory, SoftDeletes, \App\Traits\Multitenantable, \Spatie\Activitylog\Traits\LogsActivity;
    protected $primaryKey = 'id';
    public $incrementing = false;
    protected $keyType = 'string';
    protected $fillable = [
        'organization_id',
        'creator_user_id',
        'project_type_id',
        'project_code',
        'title',
        'short_title',
        'description',
        'general_objectives',
        'status',
        'start_date',
        'end_date',
    ];

    protected $casts = [
        'status' => \App\Enums\ProjectStatus::class,
        'start_date' => 'date', 
        'end_date' => 'date',
        'general_objectives' => 'array',
    ];

    protected static function boot()
    {
        parent::boot();
        static::creating(fn ($model) => $model->{$model->getKeyName()} = (string) Str::uuid());
    }

    public function getActivitylogOptions(): \Spatie\Activitylog\LogOptions
    {
        return \Spatie\Activitylog\LogOptions::defaults()
            ->logAll()
            ->logOnlyDirty()
            ->dontSubmitEmptyLogs();
    }

    public function tapActivity(\Spatie\Activitylog\Models\Activity $activity, string $eventName)
    {
        $activity->organization_id = $this->organization_id ?? auth()->user()?->organization_id;
    }

    public function creator()
    {
        return $this->belongsTo(User::class, 'creator_user_id', 'id');
    }
    public function projectType()
    {
        return $this->belongsTo(ProjectType::class, 'project_type_id', 'id');
    }
    public function contexts()
    {
        return $this->hasMany(ProjectContext::class, 'project_id', 'id');
    }
    public function documents()
    {
        return $this->hasMany(ProjectDocument::class, 'project_id', 'id');
    }
    public function logicalFramework()
    {
        return $this->hasOne(LogicalFramework::class, 'project_id', 'id');
    }
    public function budgets()
    {
        return $this->hasMany(Budget::class, 'project_id', 'id');
    }
    public function progressTrackers()
    {
        return $this->hasMany(ProgressTracker::class, 'project_id', 'id');
    }
    public function qualitativeEvaluations()
    {
        return $this->hasMany(QualitativeEvaluation::class, 'project_id', 'id');
    }

    // Autres relations et méthodes...
    public function projectContext()
    {
        return $this->hasOne(ProjectContext::class);
    }

    public function projectDocuments()
    {
        return $this->hasMany(ProjectDocument::class);
    }

    public function getAllActivities(): Collection
    {
        $activities = collect([]);
        // Eager load logical framework and its descendants to minimize queries
        $this->loadMissing('logicalFramework.specificObjectives.results.activities');

        if ($this->logicalFramework) {
            foreach ($this->logicalFramework->specificObjectives as $specificObjective) {
                foreach ($specificObjective->results as $result) {
                    $activities = $activities->merge($result->activities);
                }
            }
        }

        return $activities;
    }


    /**
     * Calculate the overall progress percentage of the project.
     * This is based on the average progress of all its activities.
     *
     * @return float
     */
    

    public function calculateProjectProgress(): float
    {
        $activities = $this->getAllActivities();

        if ($activities->isEmpty()) {
            return 0.0;
        }

        $totalProgress = $activities->sum->calculateProgress(); // magie de Laravel : sum sur méthode
        // ou : $activities->sum(fn($a) => $a->calculateProgress());

        return round($totalProgress / $activities->count(), 2);
    }
    
}