# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CICA-GPRO API",
        "description": "API Documentation for CICA-GPRO project management system. Designed for AI integration and organizational interoperability.",
        "contact": {
            "email": "support@cave-tech.dev"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Primary API Server"
        }
    ],
    "paths": {
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Liste tous les projets de l'organisation",
                "operationId": "56b787d17379644f68212ab86364ed3d",
                "responses": {
                    "200": {
                        "description": "Liste des projets récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Récupérer les détails d'un projet spécifique",
                "operationId": "968df0a94f68b518ecbdfddb4116a589",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du projet",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du projet",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projet non trouvé"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Activity": {
                "title": "Activity",
                "description": "Activity model",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "in_progress",
                            "completed",
                            "overdue"
                        ]
                    },
                    "progress": {
                        "type": "number",
                        "format": "float",
                        "example": 45.5
                    }
                },
                "type": "object"
            },
            "Project": {
                "title": "Project",
                "description": "Project model",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "organization_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string",
                        "example": "Construction d'un complexe scolaire"
                    },
                    "project_code": {
                        "type": "string",
                        "example": "PRJ-2026-001"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "proposed",
                            "approved",
                            "rejected"
                        ]
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "Resource": {
                "title": "Resource",
                "description": "Resource model",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Ciment"
                    },
                    "type": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "number"
                    },
                    "unit_cost": {
                        "type": "number",
                        "format": "float"
                    },
                    "total_cost": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Jean Dupont"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jean.dupont@example.com"
                    },
                    "organization_id": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Projects",
            "description": "Projects"
        }
    ]
}